backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python-headless
sqlalchemy
psycopg2-binary
redis
celery
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
gunicorn
